+++ lib/lift/AArch64/AArch64LiftNEONReduce.cpp
@@ -50,7 +50,8 @@ bool liftNEONReduce(AArch64Lifter &L, AArch64Lifter::LiftState &S,
       NdVar Acc = S.makeTemp(0);
       for (unsigned I = 0; I < NLanes; ++I) {
         NdVar Lane = S.makeTemp(LaneSz);
-        S.emit(NdOp::SUBBYTES, Lane, {Src, NdVar::cst(I * LaneSz, 4)});
+        S.emit(NdOp::SUBBYTES, Lane,
+               {Src, NdVar::cst(static_cast<uint64_t>(I) * LaneSz, 4)});
         NdVar Neg = S.makeTemp(LaneSz);
         S.emit(NdOp::INT_NEG2, Neg, {Lane});
         NdVar IsNeg = S.makeTemp(1);
